[add] made views directory.
